:running: structure and expose cobra cmds
This restructures the cmd directory into pkgs that follows the
hierarchy of CLI subcommands. This helps to expose the New***Command() functions
that return the cobra subcommands so they can be consumed by downstream
projects that need to reuse Kubebuilder CLI for scaffolding projects.
