+++ cmd/init/init_project.go
@@ -14,9 +14,10 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package init
 
 import (
+	"encoding/json"
 	"fmt"
 	"log"
 	"os"
@@ -27,13 +28,15 @@ import (
 
 	"github.com/spf13/cobra"
 	flag "github.com/spf13/pflag"
+	"golang.org/x/tools/go/packages"
 
 	"sigs.k8s.io/kubebuilder/cmd/util"
 	"sigs.k8s.io/kubebuilder/pkg/scaffold"
 	"sigs.k8s.io/kubebuilder/pkg/scaffold/project"
+	sutil "sigs.k8s.io/kubebuilder/pkg/scaffold/util"
 )
 
-func newInitProjectCmd() *cobra.Command {
+func NewInitProjectCmd() *cobra.Command {
 	o := projectOptions{}
 
 	initCmd := &cobra.Command{
@@ -64,6 +67,14 @@ kubebuilder init --domain example.org --license apache2 --owner "The Kubernetes 
 	o.bindCmdlineFlags(initCmd)
 
 	return initCmd
+}
+
+// module and goMod arg just enough of the output of `go mod edit -json` for our purposes
+type goMod struct {
+	Module module
+}
+type module struct {
+	Path string
 }
 
 type projectOptions struct {
@@ -222,6 +233,70 @@ func checkGoVersion(verStr string) error {
 	}
 
 	return nil
+}
+
+// findCurrentRepo attempts to determine the current repository
+// though a combination of go/packages and `go mod` commands/tricks.
+func findCurrentRepo() (string, error) {
+	// easiest case: project file already exists
+	projFile, err := sutil.LoadProjectFile("PROJECT")
+	if err == nil {
+		return projFile.Repo, nil
+	}
+
+	// next easy case: existing go module
+	path, err := findGoModulePath(false)
+	if err == nil {
+		return path, nil
+	}
+
+	// next, check if we've got a package in the current directory
+	pkgCfg := &packages.Config{
+		Mode: packages.NeedName, // name gives us path as well
+	}
+	pkgs, err := packages.Load(pkgCfg, ".")
+	// NB(directxman12): when go modules are off and we're outside GOPATH and
+	// we don't otherwise have a good guess packages.Load will fabricate a path
+	// that consists of `_/absolute/path/to/current/directory`.  We shouldn't
+	// use that when it happens.
+	if err == nil && len(pkgs) > 0 && len(pkgs[0].PkgPath) > 0 && pkgs[0].PkgPath[0] != '_' {
+		return pkgs[0].PkgPath, nil
+	}
+
+	// otherwise, try to get `go mod init` to guess for us -- it's pretty good
+	cmd := exec.Command("go", "mod", "init")
+	cmd.Env = append(cmd.Env, os.Environ()...)
+	cmd.Env = append(cmd.Env, "GO111MODULE=on" /* turn on modules just for these commands */)
+	if _, err := cmd.Output(); err != nil {
+		if exitErr, isExitErr := err.(*exec.ExitError); isExitErr {
+			err = fmt.Errorf("%s", string(exitErr.Stderr))
+		}
+		// give up, let the user figure it out
+		return "", fmt.Errorf("could not determine repository path from module data, package data, or by initializing a module: %v", err)
+	}
+	defer os.Remove("go.mod") // clean up after ourselves
+	return findGoModulePath(true)
+}
+
+// findGoModulePath finds the path of the current module, if present.
+func findGoModulePath(forceModules bool) (string, error) {
+	cmd := exec.Command("go", "mod", "edit", "-json")
+	cmd.Env = append(cmd.Env, os.Environ()...)
+	if forceModules {
+		cmd.Env = append(cmd.Env, "GO111MODULE=on" /* turn on modules just for these commands */)
+	}
+	out, err := cmd.Output()
+	if err != nil {
+		if exitErr, isExitErr := err.(*exec.ExitError); isExitErr {
+			err = fmt.Errorf("%s", string(exitErr.Stderr))
+		}
+		return "", err
+	}
+	mod := goMod{}
+	if err := json.Unmarshal(out, &mod); err != nil {
+		return "", err
+	}
+	return mod.Module.Path, nil
 }
 
 func (o *projectOptions) postScaffold() error {
