+++ cmd/main.go
@@ -17,105 +17,32 @@ limitations under the License.
 package main
 
 import (
-	"encoding/json"
-	"fmt"
 	"log"
-	"os"
-	"os/exec"
 
 	"github.com/spf13/cobra"
-	"golang.org/x/tools/go/packages"
 
+	"sigs.k8s.io/kubebuilder/cmd/alpha"
+	"sigs.k8s.io/kubebuilder/cmd/create"
+	cmdinit "sigs.k8s.io/kubebuilder/cmd/init"
+	"sigs.k8s.io/kubebuilder/cmd/update"
+	"sigs.k8s.io/kubebuilder/cmd/util"
 	"sigs.k8s.io/kubebuilder/cmd/version"
-	"sigs.k8s.io/kubebuilder/pkg/scaffold"
 )
-
-// module and goMod arg just enough of the output of `go mod edit -json` for our purposes
-type goMod struct {
-	Module module
-}
-type module struct {
-	Path string
-}
-
-// findGoModulePath finds the path of the current module, if present.
-func findGoModulePath(forceModules bool) (string, error) {
-	cmd := exec.Command("go", "mod", "edit", "-json")
-	cmd.Env = append(cmd.Env, os.Environ()...)
-	if forceModules {
-		cmd.Env = append(cmd.Env, "GO111MODULE=on" /* turn on modules just for these commands */)
-	}
-	out, err := cmd.Output()
-	if err != nil {
-		if exitErr, isExitErr := err.(*exec.ExitError); isExitErr {
-			err = fmt.Errorf("%s", string(exitErr.Stderr))
-		}
-		return "", err
-	}
-	mod := goMod{}
-	if err := json.Unmarshal(out, &mod); err != nil {
-		return "", err
-	}
-	return mod.Module.Path, nil
-}
-
-// findCurrentRepo attempts to determine the current repository
-// though a combination of go/packages and `go mod` commands/tricks.
-func findCurrentRepo() (string, error) {
-	// easiest case: project file already exists
-	projFile, err := scaffold.LoadProjectFile("PROJECT")
-	if err == nil {
-		return projFile.Repo, nil
-	}
-
-	// next easy case: existing go module
-	path, err := findGoModulePath(false)
-	if err == nil {
-		return path, nil
-	}
-
-	// next, check if we've got a package in the current directory
-	pkgCfg := &packages.Config{
-		Mode: packages.NeedName, // name gives us path as well
-	}
-	pkgs, err := packages.Load(pkgCfg, ".")
-	// NB(directxman12): when go modules are off and we're outside GOPATH and
-	// we don't otherwise have a good guess packages.Load will fabricate a path
-	// that consists of `_/absolute/path/to/current/directory`.  We shouldn't
-	// use that when it happens.
-	if err == nil && len(pkgs) > 0 && len(pkgs[0].PkgPath) > 0 && pkgs[0].PkgPath[0] != '_' {
-		return pkgs[0].PkgPath, nil
-	}
-
-	// otherwise, try to get `go mod init` to guess for us -- it's pretty good
-	cmd := exec.Command("go", "mod", "init")
-	cmd.Env = append(cmd.Env, os.Environ()...)
-	cmd.Env = append(cmd.Env, "GO111MODULE=on" /* turn on modules just for these commands */)
-	if _, err := cmd.Output(); err != nil {
-		if exitErr, isExitErr := err.(*exec.ExitError); isExitErr {
-			err = fmt.Errorf("%s", string(exitErr.Stderr))
-		}
-		// give up, let the user figure it out
-		return "", fmt.Errorf("could not determine repository path from module data, package data, or by initializing a module: %v", err)
-	}
-	defer os.Remove("go.mod") // clean up after ourselves
-	return findGoModulePath(true)
-}
 
 func main() {
 	rootCmd := defaultCommand()
 
 	rootCmd.AddCommand(
-		newInitProjectCmd(),
-		newCreateCmd(),
+		cmdinit.NewInitProjectCmd(),
+		create.NewCreateCmd(),
 		version.NewVersionCmd(),
 	)
 
-	foundProject, version := getProjectVersion()
+	foundProject, version := util.GetProjectVersion()
 	if foundProject && version == "1" {
 		rootCmd.AddCommand(
-			newAlphaCommand(),
-			newVendorUpdateCmd(),
+			alpha.NewAlphaCommand(),
+			update.NewVendorUpdateCmd(),
 		)
 	}
 
@@ -174,17 +101,4 @@ After the scaffold is written, api will run make on the project.
 			cmd.Help()
 		},
 	}
-}
-
-// getProjectVersion tries to load PROJECT file and returns if the file exist
-// and the version string
-func getProjectVersion() (bool, string) {
-	if _, err := os.Stat("PROJECT"); os.IsNotExist(err) {
-		return false, ""
-	}
-	projectInfo, err := scaffold.LoadProjectFile("PROJECT")
-	if err != nil {
-		log.Fatalf("failed to read the PROJECT file: %v", err)
-	}
-	return true, projectInfo.Version
 }
